+++ elements/upfront-image/js/uimage.js
@@ -90,6 +90,8 @@ define([
 
 			this.listenTo(Upfront.Events, 'command:layout:save', this.saveResizing);
 			this.listenTo(Upfront.Events, 'command:layout:save_as', this.saveResizing);
+			
+			this.listenTo(Upfront.Events, "preset:image:updated", this.caption_updated, this);
 
 			this.listenTo(Upfront.Events, 'upfront:layout_size:change_breakpoint', function(newMode){
 				if(newMode.id !== 'desktop') {
@@ -131,8 +133,16 @@ define([
 
 			return props[prop_name];
 		},
-		preset_updated: function() {
+		preset_updated: function(preset) {
 			this.render();
+			Upfront.Events.trigger('preset:image:updated', preset);
+		},
+		
+		caption_updated: function(preset) {
+			var currentPreset = this.model.get_property_value_by_name("preset");
+
+			//If element use updated preset re-render
+			if(currentPreset === preset) this.render();
 		},
 
 		update_colors: function () {
@@ -448,6 +458,11 @@ define([
 			props.l10n = l10n.template;
 
 			props.usingNewAppearance = props.usingNewAppearance || false;
+
+			// Clean up hardcoded caption color
+			if (props.usingNewAppearance) {
+				props.image_caption = props.image_caption.replace(/^<span style=".+?"/, '<span ');
+			}
 
 			rendered = this.imageTpl(props);
 
@@ -856,28 +871,32 @@ define([
 
 			//Wonderful stuff from here down
 			this.$('.uimage').css('height', data.elementSize.height);
-
+			
+			/*
+			 * Disable image resize on element resize, use Image Edit mode instead
+			 */
+			
 			//Resizing the stretching dimension has priority, the other dimension just alter position
-			if(data.stretch && !data.vstretch){
-				this.resizingH(img, data, true);
-				this.resizingV(img, data);
-			} else if(!data.stretch && data.vstretch){
-				this.resizingV(img, data, true);
-				this.resizingH(img, data);
-			} else {
+			//if(data.stretch && !data.vstretch){
+			//	this.resizingH(img, data, true);
+			//	this.resizingV(img, data);
+			//} else if(!data.stretch && data.vstretch){
+			//	this.resizingV(img, data, true);
+			//	this.resizingH(img, data);
+			//} else {
 				//Both stretching or not stretching, calculate ratio difference
-				ratio = data.size.width / data.size.height - data.elementSize.width / data.elementSize.height;
+			//	ratio = data.size.width / data.size.height - data.elementSize.width / data.elementSize.height;
 
 				//Depending on the difference of ratio, the resizing is made horizontally or vertically
-				if(ratio > 0 && data.stretch || ratio < 0 && ! data.stretch){
-					this.resizingV(img, data, true);
-					this.resizingH(img, data);
-				}
-				else {
-					this.resizingH(img, data, true);
-					this.resizingV(img, data);
-				}
-			}
+			//	if(ratio > 0 && data.stretch || ratio < 0 && ! data.stretch){
+			//		this.resizingV(img, data, true);
+			//		this.resizingH(img, data);
+			//	}
+			//	else {
+			//		this.resizingH(img, data, true);
+			//		this.resizingV(img, data);
+			//	}
+			//}
 
 			this.updateControls();
 			this.setupBySize();
@@ -905,7 +924,7 @@ define([
 				};
 				this.property('element_size', this.elementSize);
 				return;
-				
+
 			//} else if (this.property('quick_swap')) {
 			} else if (this.isThemeImage()) {
 				return;