+++ elements/upfront-slider/js/uslider.js
@@ -64,6 +64,9 @@ var USliderView = Upfront.Views.ObjectView.extend({
 			'click .upfront-icon-prev': 'prevSlide',
 			'click .uslider-starting-options': 'checkStartingInputClick'
 		});
+		
+		//Update slide defaults to match preset settings
+		this.updateSlideDefaults();
 
 		this.model.slideCollection = new Uslider_Slides(this.property('slides'));
 
@@ -81,6 +84,7 @@ var USliderView = Upfront.Views.ObjectView.extend({
 		});
 
 		this.listenTo(Upfront.Events, "theme_colors:update", this.update_colors, this);
+		this.listenTo(Upfront.Events, "preset:slider:updated", this.caption_updated, this);
 
 		this.listenTo(this.model, "preset:updated", this.preset_updated);
 
@@ -95,6 +99,18 @@ var USliderView = Upfront.Views.ObjectView.extend({
 
 		//Current Slide index
 		this.setCurrentSlide(0);
+	},
+	
+	updateSlideDefaults: function() {
+		primary = this.get_preset_properties().primaryStyle,
+			defaults = {
+				below: 'below',
+				over: 'bottomOver',
+				side: 'right'
+			}
+		;
+
+		Upfront.data.uslider.slideDefaults.style = defaults[primary];
 	},
 
 	get_preset_properties: function() {
@@ -114,8 +130,16 @@ var USliderView = Upfront.Views.ObjectView.extend({
 		return preset_props[key] ? preset_props[key] : false;
 	},
 
-	preset_updated: function() {
+	preset_updated: function(preset) {
 		this.render();
+		Upfront.Events.trigger('preset:slider:updated', preset);
+	},
+	
+	caption_updated: function(preset) {
+		var currentPreset = this.model.get_property_value_by_name("preset");
+
+		//If element use updated preset re-render
+		if(currentPreset === preset) this.render();
 	},
 
 	update_colors: function () {
@@ -442,17 +466,14 @@ var USliderView = Upfront.Views.ObjectView.extend({
 					primary == 'side' && _.indexOf(['right', 'left'], style) == -1)
 						slide.set('style', defaults[primary]);
 
-				if(primary == 'side' || me.lastStyle == 'side'){
 					var wrap = me.$('.uslide[rel=' + slide.id + ']').find('.uslide-image');
 					me.imageProps[slide.id] = me.calculateImageResize({width: wrap.width(), height:wrap.height()}, slide);
-				}
 			});
-			if(primary == 'side' || this.lastStyle == 'side')
-				this.setTimer();
+
+			this.setTimer();
 			this.lastStyle = primary;
 			this.onSlidesCollectionChange();
 		}
-
 	},
 	checkStartingInputClick: function(e){
 		//Hack to make the radio buttons work in the starting layout
@@ -476,6 +497,8 @@ var USliderView = Upfront.Views.ObjectView.extend({
 	},
 
 	setImageResizable: function(){
+		if(!this.model.slideCollection.length) return;
+
 		var me = this,
 			current = this.$('.upfront-default-slider-item-current'),
 			$slide = current.find('.uslide-image'),