+++ library/class_ufront_ufc.php
@@ -1,5 +1,7 @@
 <?php
 
+require_once "class_upfront_ufc_utils.php";
+
 class Upfront_UFC {
 
 	/**
@@ -19,6 +21,17 @@ class Upfront_UFC {
 
 
 	private static $_theme_color_count;
+
+	/**
+	 * @var class instance of Upfront_UFC_Utils
+	 */
+	private static $_utils;
+
+	/**
+	 * @static string ufc, used to build the ufc variable
+	 */
+	const VAR_PREFIX = "ufc";
+
 	/**
 	 * @param null $color ufc|color
 	 *
@@ -37,7 +50,7 @@ class Upfront_UFC {
 		self::$_theme_colors = is_string(self::$_theme_colors) ? json_decode( self::$_theme_colors ) : self::$_theme_colors;
 		self::$_theme_color_count = !empty(self::$_theme_colors->colors) ? count( self::$_theme_colors->colors ) : 0;
 
-		if( strpos( $color, "ufc" ) !== false ){
+		if( strpos( $color, self::VAR_PREFIX ) !== false ){
 			self::$_ufc = $color;
 		}else{
 			self::$_color = $color;
@@ -58,14 +71,14 @@ class Upfront_UFC {
 	public function get_ufc( $color = null){
 		$color = empty( $color ) ? self::$_ufc : $color;
 
-		if( strpos($color, "ufc")  !== false ) return $color;
+		if( strpos($color, self::VAR_PREFIX)  !== false ) return $color;
 
-		$color = $this->_hex2rgb( $color );
+		$color = self::utils()->hex2rgb( $color );
 
 		$theme_colors = self::$_theme_colors->colors;
 
 		foreach( $theme_colors as $key => $theme_color ){
-			if( $color === $this->_hex2rgb(  $theme_color->color ) ) return "ufc" . $key;
+			if( $color === self::utils()->hex2rgb(  $theme_color->color ) ) return self::VAR_PREFIX . $key;
 		}
 		return false;
 	}
@@ -81,7 +94,7 @@ class Upfront_UFC {
 	public function get_color( $ufc = null){
 		$ufc = empty( $ufc ) ? self::$_ufc : $ufc;
 
-		$index = (int)  str_replace("ufc", "", $ufc);
+		$index = (int)  str_replace(self::VAR_PREFIX, "", $ufc);
 
 		$theme_colors = self::$_theme_colors->colors;
 
@@ -92,70 +105,51 @@ class Upfront_UFC {
 		return false;
 	}
 
-	/**
-	 * Checks if given color is hex
-	 *
-	 * @param $color
-	 *
-	 * @return bool
-	 */
-	private function _is_hex( $color ){
-		return ctype_xdigit( $color );
-	}
 
 	/**
-	 * Checks if given color is rgb
+	 * Replaces ufc variables with actual color code
+	 * First makes sure we don't have commented ufc colors in the given string and instead we only have ufc vars
+	 * then converts the variables to actual colors
 	 *
-	 * @param $color
-	 *
-	 * @return bool
+	 * @uses Upfront_UFC_Utils::replace_commented_style_with_variable
+	 * @uses convert_variables_to_color
+	 * @param $string
+	 * @return mixed
 	 */
-	private function _is_rgb( $color ){
-		if( $this->_is_hex( $color ) ) return false;
-
-		$color = preg_replace('/\s+/', '', $color);
-		$color = explode(",", $color);
-		return count($color) === 3;
-
-	}
-
-	/**
-	 * Checks if given color is rgba
-	 *
-	 * @param $color
-	 *
-	 * @return bool
-	 */
-	private function _is_rgba( $color ){
-		return !$this->_is_rgb( $color );
-	}
-
-	private function _hex2rgb( $color ) {
-		if( !$this->_is_hex( $color ) ) return $color;
-
-		if ( $color[0] == '#' ) {
-			$color = substr( $color, 1 );
-		}
-		if ( strlen( $color ) == 6 ) {
-			list( $r, $g, $b ) = array( $color[0] . $color[1], $color[2] . $color[3], $color[4] . $color[5] );
-		} elseif ( strlen( $color ) == 3 ) {
-			list( $r, $g, $b ) = array( $color[0] . $color[0], $color[1] . $color[1], $color[2] . $color[2] );
-		} else {
-			return false;
-		}
-		$r = hexdec( $r );
-		$g = hexdec( $g );
-		$b = hexdec( $b );
-		return "(" . $r . "," . $g . "," . $b . ")";
-	}
-
 	public function process_colors( $string ){
+		$string = self::utils()->replace_commented_style_with_variable( $string );
+		return $this->convert_variables_to_color( $string );
+	}
+
+	/**
+	 * Finds and converts ufc variables to actual colors
+	 *
+	 * @param string $string css styles
+	 * @return mixed
+	 */
+	public function convert_variables_to_color( $string ){
 		$theme_colors = !empty(self::$_theme_colors->colors) ? self::$_theme_colors->colors : array();
+
 		for( $i = 0; $i < self::$_theme_color_count ; $i++ ){
-			$string = str_replace("#ufc" . $i, $theme_colors[$i]->color  , $string );
+			$string = str_replace("#" . self::VAR_PREFIX . $i, $theme_colors[$i]->color  , $string );
 		}
 
 		return $string;
 	}
+
+	/**
+	 * Returns instance of Upfront_UFC_Utils
+	 *
+	 * @return Upfront_UFC_Utils
+	 */
+	static function utils(){
+		if( self::$_utils instanceof Upfront_UFC_Utils)
+			return self::$_utils;
+
+		self::$_utils = new Upfront_UFC_Utils();
+
+		return self::$_utils;
+	}
+
 }