+++ scripts/upfront/settings/modules/migrate-preset.js
@@ -8,15 +8,15 @@ define([
 
 		initialize: function(options) {
 			this.options = options || {};
-			
+
 			var me = this;
-			
+
 			var SimpleTextField = Upfront.Views.Editor.Field.Text.extend({
 				get_field_html: function () {
 					return '';
 				}
 			});
-			
+
 			var SelectPresetField = Upfront.Views.Editor.Field.Chosen_Select.extend({
 				className: 'preset select-preset-field-overlay',
 				render: function() {
@@ -35,6 +35,9 @@ define([
 				get_value_html: function (value, index) {
 					var selected = '';
 					var currentPreset = this.get_saved_value() ? this.get_saved_value() : 'default';
+					if (!this.model.get_property_value_by_name('usingNewAppearanche')) {
+						currentPreset = '';
+					}
 					if (value.value === this.clear_preset_name(currentPreset)) selected = ' selected="selected"';
 					return ['<option value="', value.value, '"', selected, '>', value.label, '</option>'].join('');
 				},
@@ -44,22 +47,21 @@ define([
 					preset = preset.replace(/[^-a-zA-Z0-9]/, '');
 					return preset;
 				},
-				
+
 				on_change: function(e) {
 					this.trigger('change', this.get_value());
 				}
 			});
-			
+
 			this.selectPresetField = new SelectPresetField({
 					model: this.model,
 					label: '',
-					property: 'preset',
 					values: this.get_presets(),
 					change: function(value) {
 						//me.model.set_property('preset', this.get_value());
 					}
 				}),
-			
+
 			this.listenTo(this.selectPresetField, 'change', this.previewPreset);
 
 			var fields = [
@@ -72,7 +74,29 @@ define([
 						label: l10n.convert_preset_info,
 						className: 'migrate-preset-info migrate-info-icon',
 					}),
-					
+
+					new SimpleTextField({
+						model: this.model,
+						label: l10n.select_preset_info,
+						className: 'migrate-preset-info',
+					}),
+
+					this.selectPresetField,
+
+					new Upfront.Views.Editor.Field.Button({
+						model: this.model,
+						label: l10n.apply_label,
+						className: 'migrate-preset-apply',
+						compact: true,
+						on_click: function() {
+							me.trigger('upfront:presets:change', me.selectPresetField.get_value());
+						}
+					}),
+					new SimpleTextField({
+						model: this.model,
+						label: l10n.save_as_preset_button_info,
+						className: 'migrate-preset-info save-as-preset-info',
+					}),
 					new Upfront.Views.Editor.Field.Button({
 						model: this.model,
 						label: l10n.convert_style_to_preset,
@@ -82,9 +106,7 @@ define([
 							me.show_new_preset_fields();
 						}
 					}),
-					
 					//New preset fields
-					
 					new Upfront.Views.Editor.Field.Button({
 						model: this.model,
 						label: l10n.cancel_label,
@@ -94,14 +116,14 @@ define([
 							me.hide_new_preset_fields();
 						}
 					}),
-					
+
 					new Upfront.Views.Editor.Field.Text({
 						model: this.model,
 						label: '',
 						default_value: me.suggestPresetName(this.options.elementPreset),
 						className: 'new-preset-button-input',
 					}),
-					
+
 					new Upfront.Views.Editor.Field.Button({
 						model: this.model,
 						label: l10n.ok_label,
@@ -118,6 +140,10 @@ define([
 								alert(l10n.special_character_label);
 								return;
 							}
+							if (!preset_name.match(/^[A-Za-z][A-Za-z0-9 ]*$/)) {
+								alert(l10n.invalid_preset_label);
+								return;
+							}
 
 							me.trigger('upfront:presets:new', preset_name.trim());
 						}
@@ -125,76 +151,93 @@ define([
 				]
 				//End new preset fields
 			}),
-			new Upfront.Views.Editor.Settings.Item({
-				model: this.model,
-				className: 'existing-preset-module migrate-separator',
-				fields: [
-					new SimpleTextField({
-						model: this.model,
-						label: l10n.select_preset_info,
-						className: 'migrate-preset-info',
-					}),
+			// new Upfront.Views.Editor.Settings.Item({
+				// model: this.model,
+				// className: 'existing-preset-module migrate-separator',
+				// fields: [
+				// ]
+			// })
+			];
 
-					this.selectPresetField,
-					
-					new Upfront.Views.Editor.Field.Button({
-						model: this.model,
-						label: l10n.apply_label,
-						className: 'migrate-preset-apply',
-						compact: true,
-						on_click: function() {
-							me.trigger('upfront:presets:change', me.selectPresetField.get_value());
-						}
-					}),
-				]
-			})
-			]
-			
 			setTimeout(function(){
 				me.hide_new_preset_fields();
 			}, 20);
 
 			this.fields =_(fields);
 		},
-		
+
+		/**
+		 * Suggests an unique preset name, based on the preset argument
+		 *
+		 * @param {String} presetName Preset name prefix
+		 *
+		 * @return {String} Unique preset name
+		 */
 		suggestPresetName: function(presetName) {
 			var preset = this.capitalisePreset(presetName.replace(/-preset/, '')),
-				name = preset + l10n.preset;
-			
-			return name
+				raw_name = preset + l10n.preset,
+				name = raw_name,
+				counter = 1,
+				all_names = this.get_all_preset_names()
+			;
+			// Increment counter next to preset name until we get an unique one
+			while (0 < all_names.indexOf(name)) {
+				name = raw_name + counter;
+				counter++;
+			}
+			return name;
 		},
-		
+
 		capitalisePreset: function(preset) {
 			return preset.charAt(0).toUpperCase() + preset.slice(1).toLowerCase();
 		},
-		
+
 		hide_new_preset_fields: function () {
 			var me = this;
 			me.$el.find('.new-preset-button-cancel').hide();
 			me.$el.find('.new-preset-button-input').hide();
 			me.$el.find('.new-preset-button-submit').hide();
 			me.$el.find('.migrate-preset-button').show();
-			
+
 			me.$el.find('.existing-preset-overlay-layout').remove();
 		},
-		
+
 		show_new_preset_fields: function () {
 			var me = this;
 			me.$el.find('.new-preset-button-cancel').show();
 			me.$el.find('.new-preset-button-input').show();
 			me.$el.find('.new-preset-button-submit').show();
 			me.$el.find('.migrate-preset-button').hide();
-			
+
 			me.$el.find('.existing-preset-module').append('<div class="existing-preset-overlay-layout">&nbsp;</div>');
 		},
-		
+
 		previewPreset: function(value) {
 			this.trigger('upfront:presets:preview', value);
 		},
-		
-		get_presets: function () {
+
+		/**
+		 * Gets *all* preset names - legacy or otherwise
+		 *
+		 * @return {Array} Array of preset names as strings
+		 */
+		get_all_preset_names: function () {
 			var presets = [];
-			
+
+			_.each(this.options.presets.models, function(model) {
+				var name = 'undefined' === typeof model.get('name')
+					? model.get('id')
+					: model.get('name')
+				;
+				presets.push(name);
+			});
+
+			return presets;
+		},
+
+		get_presets: function () {
+			var presets = [{ label: 'No preset', value: ''}];
+
 			_.each(this.options.presets.models, function(model) {
 				if(typeof model.get('legacy') !== "undefined") {
 					return;